More misc changes, seperated out projectiles, and prevented drawing glow
